Partial match feature
A partial match will succeed if at least the first part of input
string matches: it does not fail in the case there are remaining
unconsumed characters.

Note it fails in the case the start rule fails: if the grammar
author want not to fail in this case, the start rule can be
surrounded by a "?".